refactor(mobile): rename welcome to onboard
