+++ apps/mobile/src/modules/Onboard/components/index.ts
@@ -1,0 +1,1 @@
+export * from './OnboardScreen';
