+++ apps/mobile/src/modules/Onboard/components/location-permission.png
