+++ apps/mobile/src/modules/Onboard/components/users-around-world.png
