--- apps/mobile/src/modules/Welcome/components/index.ts
@@ -1,1 +1,0 @@
-export * from './WelcomeScreen';
